refactor: modernization fixes, svelte 5 vite config, and go linting cleanup
